chore: move the leaderboard query to a seperate file for more modularity
